Fixes from Redis implementation and added async event handler for DB persistence. Updated DB to handle a JSONB column, instead of two structured tables

+++ src/main/java/com/aggregate/nexus/client/response/FinnhubResponse.java
@@ -3,8 +3,6 @@ package com.aggregate.nexus.client.response;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
-
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record FinnhubResponse(@JsonProperty("c") BigDecimal currentPrice) {
+public record FinnhubResponse(@JsonProperty("c") double currentPrice) {
 }
