+++ src/main/java/com/aggregate/nexus/domain/AggregatedQuoteResponse.java
@@ -1,12 +1,11 @@
 package com.aggregate.nexus.domain;
 
-import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Objects;
 
 public record AggregatedQuoteResponse(String symbol,
-                                      BigDecimal lastPrice,
-                                      Map<String, BigDecimal> sources) {
+                                      double lastPrice,
+                                      Map<String, Double> sources) {
     public AggregatedQuoteResponse {
         Objects.requireNonNull(symbol);
         sources = Map.copyOf(sources);
