+++ src/main/java/com/aggregate/nexus/domain/MarketData.java
@@ -1,6 +1,4 @@
 package com.aggregate.nexus.domain;
 
-import java.math.BigDecimal;
-
-public record MarketData(String sourceName, String symbol, BigDecimal currentPrice) {
+public record MarketData(String sourceName, String symbol, double currentPrice) {
 }
